refactor(component): 优化 cModule 组件中数据类型定义
- 将 writeLen 和 hexLen 的数据类型从 uint8_t 改为 uint32_t
-涉及多个 cModule 实现文件，包括 ec200u、mb26、ml307a、nb63、nb81、pt-aep_gh 和 pt-gskl
-此修改旨在提高代码的可移植性和安全性，避免潜在的数据溢出问题

+++ component/cModule/cModule-ec200u.c
@@ -1327,11 +1327,11 @@ static int32_t socketTcpSend(cModule_ProtocolTcpIpMessage_t *msg, uint32_t packM
         uint8_t *hexBuffer = calloc(1, 129);
         ASSERT(hexBuffer != NULL);
         while (cur < msg->payloadLength) {
-            uint8_t writeLen = msg->payloadLength - cur;
+            uint32_t writeLen = msg->payloadLength - cur;
             if (writeLen >= 64) {
                 writeLen = 64;
             }
-            uint8_t hexLen = klStr_hex2str((uint8_t *) msg->payload + cur, writeLen, (char *) hexBuffer);
+            uint32_t hexLen = klStr_hex2str((uint8_t *) msg->payload + cur, writeLen, (char *) hexBuffer);
 
             rilat_directWrite(&mModuleInstance.rilat.instance, (uint8_t *) hexBuffer, hexLen, 0);
             memset(hexBuffer, 0x00, 129);
